backend/pyproject.toml
[project]
name = "drumtrack-backend"
version = "0.1.0"
description = "DrumTrack backend - drum isolation & MIDI generation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "pydantic-settings>=2.7.0",
    "librosa>=0.10.2",
    "numpy>=1.26.0",
    "pretty-midi>=0.2.10",
    "yt-dlp>=2024.12.0",
    "soundfile>=0.12.1",
    "demucs>=4.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
